Santé-automat/requirements.txt
pdfplumber>=0.11,<0.12
pandas>=2.2,<3
python-dotenv>=1.0,<2
openpyxl>=3.1,<4
matplotlib>=3.9,<4
